Fix, style: 게시글 수, 팔로워 수, 팔로잉 수 출력
