+++ moamoa/src/Router/AppRoutes.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
+import LoginPage from '../Pages/LoginPage';
 import ProfileInfo from '../Pages/ProfileInfo';
 // import Landing from '../Pages/Landing';
 
@@ -9,8 +10,10 @@ export default function AppRoutes() {
       {/* <Route path='/' element={<Landing />} /> */}
 
       {/* 내 프로필 */}
-      <Route path='/user' element={<ProfileInfo />} />
+      <Route path='/profile' element={<ProfileInfo />} />
 
+      {/* 로그인 */}
+      <Route path='/login' element={<LoginPage />} />
       {/* 상품 등록 */}
       {/* <Route path="/product" element={<?? />} /> */}
